Add files via upload
Applogic IT Solutions India Pvt Ltd is based in Hyderabad, India. We bring more than a decade of proven expertise in IT Security and IT Audit Services. Importantly, we drive value for businesses in three ways. By leveraging on innovative and latest IT Security and Audit practices, we ensure high level of business efficiency and customer confidence
For more information visit the link : https://www.applogicconsulting.com/about-us/